ZKC-32 add a prepare method for customer during zk pay (to match btc approach)
