+++ src/channels_util.rs
@@ -1,5 +1,16 @@
 use super::*;
 use std::fmt::Display;
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct FundingTxInfo {
+    pub init_cust_bal: i64,
+    pub init_merch_bal: i64,
+    pub escrow_txid: FixedSizeArray32,
+    pub escrow_prevout: FixedSizeArray32,
+    pub merch_txid: FixedSizeArray32,
+    pub merch_prevout: FixedSizeArray32,
+}
+
 #[derive(Clone, Debug, PartialEq, Display, Serialize, Deserialize)]
 pub enum ProtocolStatus {
     New,
